Breaks out OpenAL from arcan_audio.c into platform/audio/openal.c
