+++ src/CMakeLists.txt
@@ -141,6 +141,7 @@ else()
 		endif()
 		include(GNUInstallDirs)
 	endif()
+	set(APLATFORM_STR "openal")
 	set(AGPPLATFORM_STR "gl21, gles2, gles3, stub")
 
 	# we can remove some of this cruft when 'buntu LTS gets ~3.0ish
@@ -161,6 +162,7 @@ else()
 	amsg("${CL_YEL}\t-DBUILD_PRESET=${CL_GRN}everything client${CL_RST}")
 	amsg("")
 	amsg("${CL_WHT}Audio/Video/Input Support:")
+	amsg("${CL_YEL}\t-DAUDIO_PLATFORM=${CL_GRN}${APLATFORM_STR}${CL_RST}")
 	amsg("${CL_YEL}\t-DVIDEO_PLATFORM=${CL_GRN}${VPLATFORM_STR}${CL_RST}")
 	amsg("${CL_YEL}\t-DAGP_PLATFORM=${CL_GRN}${AGPPLATFORM_STR}${CL_RST}")
 	amsg("")
@@ -317,36 +319,6 @@ else()
 	set(FREETYPE_DEFAULT_LIBRARIES ${FREETYPE_LIBRARIES})
 	set(FREETYPE_DEFAULT_INCLUDE_DIRS ${FREETYPE_INCLUDE_DIRS})
 
-	# need the separation here to not confuse openAL here with
-	# the version that we patch into LWA
-	if (EXISTS ${EXTERNAL_SRC_DIR}/git/openal AND STATIC_OPENAL)
-		amsg("${CL_YEL}Building OpenAL static from external/git mirror${CL_RST}")
-		ExternalProject_Add(OpenAL
-			SOURCE_DIR ${CMAKE_CURRENT_BINARY_DIR}/openal
-			BINARY_DIR ${CMAKE_CURRENT_BINARY_DIR}/openal_static
-			UPDATE_COMMAND ""
-			GIT_REPOSITORY "${EXTERNAL_SRC_DIR}/git/openal"
-			${EXTERNAL_DEFS}
-			${CMAKE_EXTERNAL_DEFS}
-			-DALSOFT_BACKEND_DSOUND=OFF
-			-DALSOFT_BACKEND_MMDEVAPI=OFF
-			-DALSOFT_BACKEND_OPENSL=OFF
-			-DALSOFT_BACKEND_PORTAUDIO=OFF
-			-DALSOFT_BACKEND_SOLARIS=OFF
-			-DALSOFT_BACKEND_SNDIO=OFF
-			-DALSOFT_BACKEND_QSA=OFF
-			-DALSOFT_BACKEND_WAVE=OFF
-			-DALSOFT_BACKEND_WINMM=OFF
-		)
-		set(OPENAL_LIBRARY
-			"${CMAKE_CURRENT_BINARY_DIR}/openal_static/libopenal.a"
-		)
-		set(OPENAL_INCLUDE_DIR "${CMAKE_CURRENT_BINARY_DIR}/include/AL")
-		list(APPEND MAIN_DEPS OpenAL)
-	else()
-		find_package(OpenAL REQUIRED QUIET)
-	endif()
-
 	#
 	# the amalgamation is already included in the external/ tree sqlite3 use is
 	# primarily settings management and very limited amount of queries, size is
@@ -364,6 +336,7 @@ else()
 
 	include(${PLATFORM_ROOT}/cmake/CMakeLists.AGP)
 	include(${PLATFORM_ROOT}/cmake/CMakeLists.Video)
+	include(${PLATFORM_ROOT}/cmake/CMakeLists.Audio)
 	set(EXTMAKE_CMD make)
 
 	# for the statically- linked external dependencies, we need to hint where
@@ -426,7 +399,6 @@ else()
 		INCLUDE_DIRS
 		${LUA_INCLUDE_DIR}
 		${LUA_INCLUDE_DIRS}
-		${OPENAL_INCLUDE_DIR}
 		${FREETYPE_DEFAULT_INCLUDE_DIRS}
 		${SQLite3_INCLUDE_DIR}
 		${CMAKE_CURRENT_SOURCE_DIR}/platform
@@ -594,6 +566,7 @@ else()
 		${AGP_SOURCES}
 	 	${PLATFORM_SOURCES}
 		${VIDEO_PLATFORM_SOURCES}
+		${AUDIO_PLATFORM_SOURCES}
 		${PLATFORM_ROOT}/${INPUT_PLATFORM}/event.c
 	)
 	add_sanitizers(arcan)
@@ -616,8 +589,8 @@ else()
 		${STDLIB}
 		${ARCAN_LIBRARIES}
 		${VIDEO_LIBRARIES}
+		${AUDIO_LIBRARIES}
 		${AGP_LIBRARIES}
-		${OPENAL_LIBRARY}
 		arcan_shmif_int
 	)
 
@@ -651,6 +624,7 @@ else()
 		set(ARCAN_DEFINITIONS "")
 		set(ARCAN_NOLWA_DEFINITIONS "")
 		set(VIDEO_PLATFORM_SOURCES "")
+		set(AUDIO_PLATFORM_SOURCES "")
 		set(INPUT_PLATFORM_SOURCES "")
 
 	# HACK: Force- remove the psep-open from the platform and switch to the open
@@ -661,6 +635,7 @@ else()
 		endif()
 
 		include(${PLATFORM_ROOT}/cmake/CMakeLists.Video)
+		include(${PLATFORM_ROOT}/cmake/CMakeLists.Audio)
 
 		add_executable(arcan_sdl
 			${SOURCES}
@@ -668,6 +643,7 @@ else()
 			${AGP_SOURCES}
 			${PLATFORM_SOURCES}
 			${VIDEO_PLATFORM_SOURCES}
+			${AUDIO_PLATFORM_SOURCES}
 			${PLATFORM_ROOT}/sdl2/event.c
 		)
 		if (MAIN_DEPS)
@@ -683,7 +659,7 @@ else()
 			FRAMESERVER_MODESTRING=\"${FRAMESERVER_MODESTRING}\"
 		)
 		target_link_libraries(arcan_sdl
-			${STDLIB} ${ARCAN_LIBRARIES} ${VIDEO_LIBRARIES} ${OPENAL_LIBRARY}
+			${STDLIB} ${ARCAN_LIBRARIES} ${VIDEO_LIBRARIES} ${AUDIO_LIBRARIES}
 			${AGP_LIBRARIES} arcan_shmif_int
 		)
 		list(APPEND BIN_INSTALL arcan_sdl)
@@ -713,6 +689,7 @@ else()
 			${AGP_SOURCES}
 			${PLATFORM_SOURCES}
 			${VIDEO_PLATFORM_SOURCES}
+			${AUDIO_PLATFORM_SOURCES}
 			${PLATFORM_ROOT}/headless/event.c
 		)
 		if (MAIN_DEPS)
@@ -727,7 +704,7 @@ else()
 			ARCAN_BUILDVERSION=\"${SOURCE_TAG}-${PLATFORM_BUILDTAG}-${CMAKE_SYSTEM_NAME}\"
 		)
 		target_link_libraries(arcan_headless
-			${STDLIB} ${ARCAN_LIBRARIES} ${VIDEO_LIBRARIES} ${OPENAL_LIBRARY}
+			${STDLIB} ${ARCAN_LIBRARIES} ${VIDEO_LIBRARIES} ${AUDIO_LIBRARIES}
 			${AGP_LIBRARIES} arcan_shmif_int
 		)
 		list(APPEND BIN_INSTALL arcan_headless)
